backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.1,<14.0
python-multipart==0.0.6

# OpenAI
openai==1.3.5

# Hume AI
hume==0.12.1

# Google Cloud Speech-to-Text
google-cloud-speech==2.27.0

# Environment and configuration
python-dotenv==1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.25.2
tenacity==8.2.3

# Visualization (for development/analysis)
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
